Updated contact name data type in db schemas

+++ db/migrate/20180306183549_remove_log_domain_statuses.rb
@@ -1,5 +1,5 @@
 class RemoveLogDomainStatuses < ActiveRecord::Migration[6.0]
   def change
-    drop_table :log_domain_statuses
+    drop_table :log_domain_statuses if table_exists?(:log_domain_statuses)
   end
 end
